update bug change mode
