+++ apps/chat/src/app/layouts/ChannelLayout.tsx
@@ -18,7 +18,7 @@ const ChannelLayout = () => {
 
 	return (
 		<div
-			className={`flex flex-col flex-1 shrink min-w-0 bg-bgSecondary h-[100%] overflow-visible ${currentChannel?.type === ChannelType.CHANNEL_TYPE_VOICE ? 'group' : ''}`}
+			className={`flex flex-col flex-1 shrink min-w-0 dark:bg-bgSecondary bg-bgLightModeSecond h-[100%] overflow-visible ${currentChannel?.type === ChannelType.CHANNEL_TYPE_VOICE ? 'group' : ''}`}
 		>
 			<div className="flex h-heightWithoutTopBar flex-row">
 				<Outlet />
