+++ apps/chat/src/index.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html lang="en" class="dark">
 	<head>
-		<meta charset="utf-8" class="dark"/>
+		<meta charset="utf-8" class=""/>
 		<title>Mezon</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<link rel="icon" type="image/x-icon" href="assets/favicon.ico" />
