+++ libs/components/src/lib/components/NavLink/index.tsx
@@ -24,7 +24,7 @@ function NavLinkComponent({ active, children, clanName}: NavLinkProps) {
 					<div
 						className={`${
 							active
-								? 'rounded-xl bg-brand text-white bg-bgTertiary'
+								? 'rounded-xl bg-brand text-white dark:bg-bgTertiary bg-bgLightModeButton'
 								: 'text-gray-100 group-hover:bg-brand group-hover:text-white group-hover:rounded-xl rounded-3xl dark:group-hover:bg-bgTertiary group-hover:bg-bgLightModeButton'
 						} flex items-center justify-center w-12 h-12 transition-all duration-200 overflow-hidden `}
 					>
